// src/types/player-analytics.type.ts
export type PlayerAnalytics = {
	acs: {
		average: number
		highest: number
		lowest: number
		variance: number
	}
	hs: {
		average: number
		highest: number
		lowest: number
		variance: number
	}
	kd: {
		average: number
		highest: number
		lowest: number
		variance: number
	}
	kda: {
		average: number
		highest: number
		lowest: number
		variance: number
	}
	adr: {
		average: number
		highest: number
		lowest: number
		variance: number
	}
	dd: {
		average: number
		highest: number
		lowest: number
		variance: number
	}
}